fix(state): satisfy SQLite identity CI gates
