+++ src/infra/state-migrations.device-identity.ts
@@ -8,17 +8,11 @@ import {
   runOpenClawStateWriteTransaction,
 } from "../state/openclaw-state-db.js";
 import {
-  validateStoredDeviceIdentity,
-  type DeviceIdentity,
-  type StoredDeviceIdentity,
-} from "./device-identity-store.js";
-import {
-  decodeCanonicalBase64OrBase64Url,
-  deriveEd25519PrivateKeyRaw,
-  deriveEd25519PublicKeyRaw,
-  ed25519PrivateKeyPemFromRaw,
-  ed25519PublicKeyPemFromRaw,
-} from "./ed25519-signature.js";
+  normalizeLegacyDeviceIdentity,
+  type NormalizedLegacyDeviceIdentity,
+} from "./device-identity-legacy.js";
+import { validateStoredDeviceIdentity, type DeviceIdentity } from "./device-identity-store.js";
+import { deriveEd25519PrivateKeyRaw, deriveEd25519PublicKeyRaw } from "./ed25519-signature.js";
 import { formatErrorMessage } from "./errors.js";
 import { acquireGatewayLock, GatewayLockError } from "./gateway-lock.js";
 import {
@@ -40,49 +34,8 @@ const MIGRATION_LOCK_POLL_INTERVAL_MS = 25;
 const MAX_LEGACY_IDENTITY_BYTES = 128 * 1024;
 const utf8Decoder = new TextDecoder("utf-8", { fatal: true });
 
-export type NormalizedLegacyDeviceIdentity = StoredDeviceIdentity;
-
-function fingerprintPublicKey(publicKeyPem: string): string {
-  return createHash("sha256").update(deriveEd25519PublicKeyRaw(publicKeyPem)).digest("hex");
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === "object" && !Array.isArray(value);
-}
-
 function isValidCreatedAtMs(value: unknown): value is number {
   return typeof value === "number" && Number.isSafeInteger(value) && value >= 0;
-}
-
-function normalizeLegacyCreatedAtMs(value: unknown): number {
-  // Shipped file readers accepted valid keypairs even when this metadata was
-  // missing or invalid. Doctor preserves that upgrade path without weakening SQLite.
-  return isValidCreatedAtMs(value) ? value : Date.now();
-}
-
-function normalizeLegacyKeyPair(params: {
-  createdAtMs: number;
-  privateKeyPem: string;
-  publicKeyPem: string;
-}): NormalizedLegacyDeviceIdentity | null {
-  try {
-    const publicKeyRaw = deriveEd25519PublicKeyRaw(params.publicKeyPem);
-    const privateKeyRaw = deriveEd25519PrivateKeyRaw(params.privateKeyPem);
-    const publicKeyPem = ed25519PublicKeyPemFromRaw(publicKeyRaw);
-    const privateKeyPem = ed25519PrivateKeyPemFromRaw(privateKeyRaw);
-    // Legacy deviceId was derived metadata. Preserve the authoritative key bytes and
-    // recompute the fingerprint so stale metadata never rotates a shipped identity.
-    const normalized = {
-      deviceId: fingerprintPublicKey(publicKeyPem),
-      publicKeyPem,
-      privateKeyPem,
-      createdAtMs: params.createdAtMs,
-    };
-    validateStoredDeviceIdentity(normalized);
-    return normalized;
-  } catch {
-    return null;
-  }
 }
 
 function deviceIdentityKeyMaterialMatches(left: DeviceIdentity, right: DeviceIdentity): boolean {
@@ -98,45 +51,6 @@ function deviceIdentityKeyMaterialMatches(left: DeviceIdentity, right: DeviceIde
   } catch {
     return false;
   }
-}
-
-/** Normalize a retired Node PEM or Swift raw-key payload for Doctor import. */
-export function normalizeLegacyDeviceIdentity(
-  value: unknown,
-): NormalizedLegacyDeviceIdentity | null {
-  if (
-    isRecord(value) &&
-    value.version === 1 &&
-    typeof value.deviceId === "string" &&
-    typeof value.publicKeyPem === "string" &&
-    typeof value.privateKeyPem === "string"
-  ) {
-    return normalizeLegacyKeyPair({
-      createdAtMs: normalizeLegacyCreatedAtMs(value.createdAtMs),
-      privateKeyPem: value.privateKeyPem,
-      publicKeyPem: value.publicKeyPem,
-    });
-  }
-  if (
-    isRecord(value) &&
-    !("version" in value) &&
-    typeof value.deviceId === "string" &&
-    typeof value.publicKey === "string" &&
-    typeof value.privateKey === "string"
-  ) {
-    try {
-      const publicKeyRaw = decodeCanonicalBase64OrBase64Url(value.publicKey);
-      const privateKeyRaw = decodeCanonicalBase64OrBase64Url(value.privateKey);
-      return normalizeLegacyKeyPair({
-        createdAtMs: normalizeLegacyCreatedAtMs(value.createdAtMs),
-        privateKeyPem: ed25519PrivateKeyPemFromRaw(privateKeyRaw),
-        publicKeyPem: ed25519PublicKeyPemFromRaw(publicKeyRaw),
-      });
-    } catch {
-      return null;
-    }
-  }
-  return null;
 }
 
 type DeviceIdentityMigrationDatabase = Pick<
